+++ Arena/ArenaLobbyData.cs
@@ -56,6 +56,8 @@ namespace RainMeadow
             [OnlineField(group = "arenaSetup")]
             public int watcherCamoLimit;
             [OnlineField(group = "arenaSetup")]
+            public int watcherRippleLevel;
+            [OnlineField(group = "arenaSetup")]
             public bool sainot;
             [OnlineField(group = "arenaSetup")]
             public bool painCatEgg;
@@ -135,6 +137,7 @@ namespace RainMeadow
                 sainot = arena.sainot;
                 saintAscendanceTimer = arena.arenaSaintAscendanceTimer;
                 watcherCamoLimit = arena.watcherCamoTimer;
+                watcherRippleLevel = arena.watcherRippleLevel;
                 currentGameMode = arena.currentGameMode;
                 currentLevel = arena.currentLevel;
                 totalLevels = arena.totalLevelCount;
@@ -186,7 +189,7 @@ namespace RainMeadow
                 (lobby.gameMode as ArenaOnlineGameMode).sainot = sainot;
                 (lobby.gameMode as ArenaOnlineGameMode).arenaSaintAscendanceTimer = saintAscendanceTimer;
                 (lobby.gameMode as ArenaOnlineGameMode).watcherCamoTimer = watcherCamoLimit;
-
+                (lobby.gameMode as ArenaOnlineGameMode).watcherRippleLevel = watcherRippleLevel;
                 (lobby.gameMode as ArenaOnlineGameMode).currentGameMode = currentGameMode;
                 (lobby.gameMode as ArenaOnlineGameMode).currentLevel = currentLevel;
                 (lobby.gameMode as ArenaOnlineGameMode).totalLevelCount = totalLevels;