+++ Game/MeadowRemixOptions.cs
@@ -38,6 +38,7 @@ public class RainMeadowOptions : OptionInterface
     public readonly Configurable<bool> ArenaItemSteal;
     public readonly Configurable<bool> WeaponCollisionFix;
     public readonly Configurable<bool> EnablePiggyBack;
+    public readonly Configurable<int> ArenaWatcherRippleLevel;
 
     public readonly Configurable<Color> MartyrTeamColor, OutlawsTeamColor, DragonSlayersTeamColor, ChieftainTeamColor;
     public readonly Configurable<string> MartyrTeamName;
@@ -57,6 +58,8 @@ public class RainMeadowOptions : OptionInterface
     public readonly Configurable<bool> StopMovementWhileSpectateOverlayActive;
 
     public readonly Configurable<bool> DevNightskySkin;
+
+    public readonly Configurable<bool> EnableAchievementsOnline;
 
     public readonly Configurable<IntroRoll> PickedIntroRoll;
 
@@ -105,6 +108,7 @@ public class RainMeadowOptions : OptionInterface
         PainCatLizard = config.Bind("PainCatLizard", true);
         BlockMaul = config.Bind("BlockMaul", false);
         BlockArtiStun = config.Bind("BlockArtiStun", false);
+        ArenaWatcherRippleLevel = config.Bind("ArenaWatcherRippleLevel", 1);
 
 
         MartyrTeamColor = config.Bind("MartyrTeamColor", new Color(1, 0.49f, 0.49f));
@@ -144,6 +148,8 @@ public class RainMeadowOptions : OptionInterface
         ChatBgOpacity = config.Bind("ChatBgOpacity", 0.2f);
 
         DevNightskySkin = config.Bind("DevNightskySkin", false);
+
+        EnableAchievementsOnline = config.Bind("EnableAchievementsOnline", false);
     }
 
     public override void Initialize()
@@ -336,8 +342,8 @@ public class RainMeadowOptions : OptionInterface
             };
             opTab.AddItems(GeneralUIArrPlayerOptions);
 
-            OnlineStorySettings = new UIelement[11]
-           {    new OpLabel(10f, 550f, Translate("Story"), bigText: true),
+            OnlineStorySettings =
+            [   new OpLabel(10f, 550f, Translate("Story"), bigText: true),
 
                 new OpLabel(10f, 500, Translate("Ready to shelter/gate"), bigText: false),
 
@@ -364,8 +370,12 @@ public class RainMeadowOptions : OptionInterface
                 new OpLabel(40f, 260, RWCustom.Custom.ReplaceLineDelimeters(Translate("Steal items from other players in Story mode")))
                 {
                     verticalAlignment = OpLabel.LabelVAlignment.Center
-                }
-           };
+                },
+
+                new OpLabel(new Vector2(40, 170), new(25, 25), Custom.ReplaceLineDelimeters(Translate("Gain achievements online")), FLabelAlignment.Left),
+                new OpCheckBox(EnableAchievementsOnline, new Vector2(10, 170)),
+
+            ];
             storyTab.AddItems(OnlineStorySettings);
 
             OpLabel arenaSpoilerLabel, slugpupHellBackgroundLabel;