+++ Game/RainMeadow.PlayerHooks.cs
@@ -93,7 +93,8 @@ public partial class RainMeadow
             WatcherOverrideRippleLevel = false;
         };
         On.Player.CamoUpdate += Player_CamoUpdate;
-
+        On.Player.ToggleCamo += Player_ToggleCamo;
+        IL.Player.TransitionRippleUpdate += Player_TransitionRippleUpdate;
 
 
         On.Player.SetMalnourished += Player_SetMalnourished;
@@ -196,7 +197,35 @@ public partial class RainMeadow
         }
         return orig(self);
     }
-
+    private void Player_ToggleCamo(On.Player.orig_ToggleCamo orig, Player self)
+    {
+        bool lastRippleState = self.room.game.cameras[0].lastRippleState;
+        orig(self);
+        if (!self.IsLocal(out _))
+            self.room.game.cameras[0].lastRippleState = lastRippleState;
+    }
+    private void Player_TransitionRippleUpdate(ILContext il)
+    {
+        try
+        {
+            ILCursor c = new(il);
+            ILLabel label = null;
+            c.GotoNext(x => x.MatchLdarg(0), x => x.MatchLdfld<Player>(nameof(Player.isCamo)), x => x.MatchBrfalse(out label));
+            c.Emit(OpCodes.Ldarg_0);
+            c.EmitDelegate(delegate (Player self)
+            {
+                return self.IsLocal(out _);
+            });
+            //if ripplespace avaliable, plays exiting ripple space animation when player isnt local.
+            //its stupid how roomcamera doesnt use FullScreen Ripple and uses the transition ripple when player transitions to ripple screen
+            //fsRipple only is used for the next rooms you enter if camera.lastRippleState is true. ToggleCamo hook prevents it being set true by non-local player
+            c.Emit(OpCodes.Brfalse, label);
+        }
+        catch (Exception ex)
+        {
+            Error(ex);
+        }
+    }
     private void Player_CamoUpdate(On.Player.orig_CamoUpdate orig, Player self)
     {
         if (self.isCamo && (!self.Consious || self.warpExhausionTime > 0))
@@ -352,7 +381,7 @@ public partial class RainMeadow
         {
             if (slugcatStatsPerPlayer.TryGetValue(self, out var stats))
             {
-                return stats.malnourished;
+                return stats.malnourished || stats.malnourishedByCreature;
             }
         }
 
@@ -362,13 +391,15 @@ public partial class RainMeadow
 
 
 
-    void Player_SetMalnourished(On.Player.orig_SetMalnourished orig, Player self, bool m)
+    void Player_SetMalnourished(On.Player.orig_SetMalnourished orig, Player self, bool m, bool malnourishedByCreature)
     {
-        orig(self, m);
+        orig(self, m, malnourishedByCreature);
         if (OnlineManager.lobby != null && !self.isNPC)
         {
             slugcatStatsPerPlayer.Remove(self);
-            slugcatStatsPerPlayer.Add(self, new SlugcatStats(self.SlugCatClass, m));
+            var newStats = new SlugcatStats(self.SlugCatClass, m);
+            newStats.malnourishedByCreature = malnourishedByCreature;
+            slugcatStatsPerPlayer.Add(self, newStats);
         }
     }
     private void SlugOnBack_GraphicsModuleUpdated(ILContext ctx)
@@ -430,6 +461,34 @@ public partial class RainMeadow
     private void Player_UpdateMSC(On.Player.orig_UpdateMSC orig, Player self)
     {
         orig(self);
+        if (isArenaMode(out var arena))
+        {
+            if (self.SlugCatClass == MoreSlugcats.MoreSlugcatsEnums.SlugcatStatsName.Saint)
+            {
+                if (!arena.sainot)
+                {
+                    if (!arena.countdownInitiatedHoldFire)
+                    {
+
+                        if (self.monkAscension == false && self.godTimer != self.maxGodTime)
+                        {
+
+                            if (self.tongue.mode == Player.Tongue.Mode.Retracted && (self.input[0].x != 0 || self.input[0].y != 0 || self.input[0].jmp))
+                            {
+                                self.godTimer += 0.8f;
+                            }
+                            else
+                            {
+                                self.godTimer -= 0.8f;
+                            }
+                        }
+
+                    }
+
+                }
+
+            }
+        }
         if (OnlineManager.lobby != null && HasSlugcatClassOnBack(self, MoreSlugcats.MoreSlugcatsEnums.SlugcatStatsName.Saint, out Player saint_player))
         {
             if (self.tongue is not null && self.tongue.Attached)
@@ -992,8 +1051,6 @@ public partial class RainMeadow
                 {
                     InputOverride.HoldFire(self);
                 }
-
-                ArenaHelpers.OverideSlugcatClassAbilities(self, arena);
             }
 
             if (ModManager.MSC)
@@ -1650,7 +1707,7 @@ public partial class RainMeadow
             }
 
             // Allow glow for any non-watcher in watcher campaign
-            if (ModManager.Watcher && self.room.game.session is StoryGameSession storyGameSession && self.rippleLevel > 0f && self.room != null && self.AI == null)
+            if (isStoryMode(out var storyGameMode) && ModManager.Watcher && self.room.game.session is StoryGameSession storyGameSession && storyGameMode.currentCampaign == Watcher.WatcherEnums.SlugcatStatsName.Watcher && self.rippleLevel > 0f && self.room != null && self.AI == null)
             {
                 storyGameSession.saveState.theGlow = true;
                 self.glowing = storyGameSession.saveState.theGlow || self.room.game.setupValues.playerGlowing;