+++ GameModes/ArenaCompetitiveGameMode.cs
@@ -88,6 +88,7 @@ namespace RainMeadow
 
         public int arenaSaintAscendanceTimer = RainMeadow.rainMeadowOptions.ArenaSaintAscendanceTimer.Value;
         public int watcherCamoTimer = RainMeadow.rainMeadowOptions.ArenaWatcherCamoTimer.Value;
+        public int watcherRippleLevel = RainMeadow.rainMeadowOptions.ArenaWatcherRippleLevel.Value;
 
 
         public ArenaClientSettings arenaClientSettings;
@@ -635,7 +636,6 @@ namespace RainMeadow
         static HashSet<AbstractPhysicalObject.AbstractObjectType> blockList = new()
         {
             AbstractPhysicalObject.AbstractObjectType.BlinkingFlower,
-            AbstractPhysicalObject.AbstractObjectType.SporePlant,
             AbstractPhysicalObject.AbstractObjectType.AttachedBee
 
         };
@@ -677,10 +677,6 @@ namespace RainMeadow
 
         public override bool AllowedInMode(PlacedObject item)
         {
-            if (item.type == PlacedObject.Type.SporePlant)
-            {
-                return false;
-            }
             if (item.type == PlacedObject.Type.StuckDaddy)
             {
                 return OnlineManager.lobby.isOwner;