+++ Menu/Pages/ArenaSlugcatSelectPage.cs
@@ -13,7 +13,7 @@ namespace RainMeadow.UI.Pages;
 public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.SelectOneButtonOwner
 {
     public SimplerButton backButton;
-    public MenuLabel slugcatNameLabel, descriptionLabel, readyWarningLabel;
+    public MenuLabel slugcatNameLabel, descriptionLabel, readyWarningLabel, chooseYourSlugcatLabel;
     public EventfulSelectOneButton[] slugcatSelectButtons;
     public MenuIllustration[] slugcatIllustrations;
     public List<SlugcatStats.Name[]> slugcatSelectNamePages;
@@ -26,7 +26,8 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
     public string painCatName, painCatDescription;
     public string defaultReadyWarningText = "You have been unreadied. Switch back to re-ready yourself automatically";
 
-    public int maxScugsPerRow = 6;
+    public static int maxScugsPerRow = 6; //You can change this
+    public static int maxScugsPerPage = maxScugsPerRow * 2; //But don't touch this
     public ArenaOnlineGameMode? Arena => OnlineManager.lobby?.gameMode as ArenaOnlineGameMode;
     public ArenaOnlineLobbyMenu? ArenaMenu => menu as ArenaOnlineLobbyMenu;
 
@@ -44,21 +45,23 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
         slugcatSelectButtons = new EventfulSelectOneButton[ArenaHelpers.selectableSlugcats.Count];
         slugcatIllustrations = new MenuIllustration[ArenaHelpers.selectableSlugcats.Count];
         slugcatSelectNamePages = new List<SlugcatStats.Name[]>();
-        for (int i=0; i<Mathf.Ceil(ArenaHelpers.selectableSlugcats.Count / (2*maxScugsPerRow))+1; i++)
+
+        int RequiredPages = (int)Mathf.Ceil((float)ArenaHelpers.selectableSlugcats.Count / maxScugsPerPage);
+        RainMeadow.Debug("Need " + RequiredPages + " pages for " + ArenaHelpers.selectableSlugcats.Count + " scugs (@" + maxScugsPerPage + "/page):");
+        for (int i=0; i< RequiredPages; i++)
         {
-            slugcatSelectNamePages.Add(new SlugcatStats.Name[Math.Min(2*maxScugsPerRow, ArenaHelpers.selectableSlugcats.Count - (2*maxScugsPerRow*i))]);
-            RainMeadow.Debug("Page " + i + " should hold " + slugcatSelectNamePages[i].Length + " scugs:");
+            slugcatSelectNamePages.Add(new SlugcatStats.Name[Math.Min(maxScugsPerPage, ArenaHelpers.selectableSlugcats.Count - (i * maxScugsPerPage))]);
+            RainMeadow.Debug("  Page " + i + " should hold " + slugcatSelectNamePages[i].Length + " scugs:");
             for (int j=0; j<slugcatSelectNamePages[i].Length; j++)
             {
-                slugcatSelectNamePages[i][j] = ArenaHelpers.selectableSlugcats[(i*(2*maxScugsPerRow))+j];
+                slugcatSelectNamePages[i][j] = ArenaHelpers.selectableSlugcats[(i*(maxScugsPerPage))+j];
                 RainMeadow.Debug("    " + j + ": " + slugcatSelectNamePages[i][j]);
             }
         }
-        SwitchSlugcatTabBy(0);
 
         painCatDescription = ModManager.MSC ? GetPainCatDescription() : "";
 
-        MenuLabel chooseYourSlugcatLabel = new(menu, this, menu.Translate("CHOOSE YOUR SLUGCAT"), new Vector2(680f, 575f), default, true);
+        chooseYourSlugcatLabel = new(menu, this, menu.Translate("CHOOSE YOUR SLUGCAT"), new Vector2(680f, 575f), default, true);
         chooseYourSlugcatLabel.label.color = new Color(0.5f, 0.5f, 0.5f);
         chooseYourSlugcatLabel.label.shader = menu.manager.rainWorld.Shaders["MenuTextCustom"];
 
@@ -84,6 +87,8 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
             descriptionGradientsPos[i] = new Vector2(680f, i > 1 ? (OnlineManager.lobby.isOwner ? 240f : 280f) : 125f);
             Container.AddChild(descriptionGradients[i]);
         }
+
+        SwitchSlugcatTabBy(0); //Switch by 0 to initialize.
 
         this.SafeAddSubobjects(backButton, prevButton, nextButton, chooseYourSlugcatLabel, readyWarningLabel, slugcatNameLabel, descriptionLabel);
         if (ArenaMenu != null)
@@ -118,7 +123,7 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
             else
                 buttonPos = i < currentButtonsInTopRow ? new Vector2(currentTopRowStartingXPos + 110f * i, 450f) : new Vector2(currentBottomRowStartingXPos + 110f * (i - currentButtonsInTopRow), 340f);
 
-            EventfulSelectOneButton btn = new(menu, this, "", "scug select", buttonPos, new Vector2(100f, 100f), slugcatSelectButtons, i + (currentSlugcatSelectPage*2*maxScugsPerRow));
+            EventfulSelectOneButton btn = new(menu, this, "", "scug select", buttonPos, new Vector2(100f, 100f), slugcatSelectButtons, i + (currentSlugcatSelectPage * maxScugsPerPage));
             SlugcatStats.Name slugcat = slugcatSelectNamePages[currentSlugcatSelectPage][i];
             string portraitFileString = ModManager.MSC && slugcat == MoreSlugcatsEnums.SlugcatStatsName.Sofanthiel ? SlugcatColorableButton.GetFileForSlugcatIndex(slugcat, painCatIndex, randomizeSofSlugcatPortrait: false) : SlugcatColorableButton.GetFileForSlugcat(slugcat, false);
             slugcatIllustrations[i] = new(menu, btn, "", portraitFileString, btn.size / 2, false, true);
@@ -164,6 +169,11 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
             Extensions.TryMassBind(TopRowElements, backButton, top: true);
             Extensions.TryMassBind(BottomRowElements, backButton, bottom: true);
         }
+
+        if (slugcatSelectNamePages.Count > 1)
+        {
+            chooseYourSlugcatLabel.text = menu.Translate("CHOOSE YOUR SLUGCAT") + " (" + (currentSlugcatSelectPage + 1) + "/" + slugcatSelectNamePages.Count + ")";
+        }
     }
 
     public void CreateArrowButtons()
@@ -189,7 +199,7 @@ public class ArenaSlugcatSelectPage : PositionedMenuObject, SelectOneButton.Sele
         nextButton.OnClick += _ => SwitchSlugcatTabBy(1);
         prevButton.OnClick += _ => menu.PlaySound(SoundID.MENU_Button_Standard_Button_Pressed);
         nextButton.OnClick += _ => menu.PlaySound(SoundID.MENU_Button_Standard_Button_Pressed);
-        if (ArenaHelpers.selectableSlugcats.Count <= 2 * maxScugsPerRow)
+        if (ArenaHelpers.selectableSlugcats.Count <= maxScugsPerPage)
         {
             prevButton.buttonBehav.greyedOut = true;
             nextButton.buttonBehav.greyedOut = true;